intelligent-report-system/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
polars==0.19.19

# NLP and Text Processing
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Topic Modeling
bertopic==0.16.0
umap-learn==0.5.5
hdbscan==0.8.33

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# LLM Integration (Open Source)
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6
httpx==0.25.2

# Vector Database
chromadb==0.4.18
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# API Framework
fastapi==0.105.0
uvicorn==0.25.0
python-multipart==0.0.6

# Utilities
requests==2.31.0
python-json-logger==2.0.7
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0